「960_修了確認_データストア４」までとりあえず終了。その他ファイル追加
