+++ dotinstall/javascript/2023011/index.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+
+<html lang="ja">
+<head>
+  <meta charset="UTF-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Document</title>
+</head>
+<body>
+  <button>Run</button>
+
+  <h1 id="target">ドットインストール</h1>
+  <p>hello</p>
+  <p>hello</p>
+  <p>hello</p>
+  <p>hello</p>
+
+  <script src="main.js"></script>
+</body>
+</html>
