+++ udemy/ガチで学びたい人のためのJavaScriptメカニズム/230310_100_様々なオブジェクト/950_修了確認_データストア３/start/main.js
@@ -2,7 +2,7 @@
  * 問題：
  * サーバー上に配置してあるJSONファイルを取得して、
  * オブジェクトにセットしましょう。
- * 
+ *
  * ただし、localStorageにすでにデータが保存されている場合
  * にはサーバーへのデータ取得は行わず、localStorageに
  * 登録されているJSONからオブジェクトを復元してください。
@@ -12,6 +12,16 @@ const p = Promise.resolve();
 let _dirty;
 
 class DataSource {
+	static async get(KEY) {
+		return DataSource.getLocal(KEY)
+		|| await DataSource.getRemote(KEY);
+	}
+
+	static async set(KEY, target) {
+		return DataSource.setLocal(KEY , target)
+		|| await DataSource.setRemote(KEY , target);
+	}
+
 	static getLocal(KEY) {
 		console.log('get from local');
 		const result = localStorage.getItem(KEY);
@@ -23,31 +33,43 @@ class DataSource {
 		const json = JSON.stringify(target);
 		localStorage.setItem(KEY, json);
 	}
+
+	static async getRemote(KEY) {
+		console.log('get from remote');
+		const response = await fetch(`../json/${KEY}.json`);
+		const json = await response.json();
+		return json;
+	}
+
+	static async setRemote(KEY, target) {
+		console.log(`set to remote: + ${KEY} + , ${target}`);
+	}
 }
 
-const targetObj = DataSource.getLocal(KEY) || {};
+(async () => {
+	const targetObj = await DataSource.get(KEY);
 
-const pxy = new Proxy(targetObj, {
-	set(target, prop, value, receiver) {
-		_dirty = true;
+	const pxy = new Proxy(targetObj, {
+		set(target, prop, value, receiver) {
+			_dirty = true;
 
-		const result = Reflect.set(target, prop, value, receiver);
+			const result = Reflect.set(target, prop, value, receiver);
 
-		p.then(() => {
-			if(_dirty) {
-				console.log('** update data **');
-				_dirty = false;
-				DataSource.setLocal(KEY, target);
-			}
-		});
-		
-		
-		return result;
-	}
-});
+			p.then(() => {
+				if(_dirty) {
+					console.log('** update data **');
+					_dirty = false;
+					DataSource.set(KEY, target);
+				}
+			});
 
-console.log('init', pxy);
-pxy.name = 'Tom';
-console.log('change', pxy);
-pxy.name = 'Tim';
-console.log('change2', pxy);
+			return result;
+		}
+	});
+
+	console.log('init', pxy);
+	pxy.name = 'Tom';
+	console.log('change', pxy);
+	pxy.name = 'Tim';
+	console.log('change2', pxy);
+})();
