+++ udemy/ガチで学びたい人のためのJavaScriptメカニズム/230310_100_様々なオブジェクト/960_修了確認_データストア４/start/index.html
@@ -20,11 +20,11 @@
         <a href="../../" style="margin-right: 10px;">
             セクションへ
         </a>
-    
+
         <a href="../../../">
             コースへ
         </a>
     </div>
-    <script src="main.js"></script>
+    <script type="module" src="main.js"></script>
 </body>
 </html>
